Changed the imports : ) Will add the functions that use them later. Each tree is stored in a file as the following:   set-length Nonce used in encryption & decryption   hash of original file, used to verify contents haven't been changed since time of encryption   encrypted JSON representation of Tree
When the file is decrypted, the key is looked up, a new cipher object generated using the first x number of characters as the Nonce (the length has yet to be decided, but will be fixed. Probably 16 or 32 bits.) The hash is stripped from the front of the file, and saved in a variable. The remainder of the file is decrypted, hashed for verification, then transferred into the cache for future use. 

The above has been coded in encryption.py, which I've failed to commit for some reason : /