Finished simplistic handling of all MCP messages. Closes #27.
Stuff doesn't route to callbacks the way they ought to, but that's a job for
another ticket entirely. As is the modularization of the vertical integration
test. But basic support for all message types is in place now, which is really
cool, and sets up perfectly for the road ahead.